파일 시스템 구현

+++ src/main/java/com/example/board/Repository/FileRepository.java
@@ -3,5 +3,5 @@ package com.example.board.Repository;
 import com.example.board.domain.AttachedFile;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface FileRepository extends JpaRepository<AttachedFile, Long> {
+public interface FileRepository extends JpaRepository<AttachedFile, Integer> {
 }
